Add new methods, tests, and errors to OnlineSimApi
This commit adds four new methods to the IOnlineSimApi protocol, namely:

1. getNumberForward to get a forward number with a specific configuration.
2. setStatus to set the status of a number, which is needed when using the forward number feature.
3. getOperationsHistory to retrieve the history of all the operations performed in the account.
4. getActivationStatus to retrieve the activation status of a phone number.
The implementation for these methods has been added to the OnlineSimApi class. Additionally, tests have been added for all the methods, which make use of MockURLProtocol to simulate HTTP requests.

Finally, the endpoints have been updated to include the new methods, and the functionality has been checked in the application with real tests.